requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.0
psycopg2-binary==2.9.7
pytesseract==0.3.10
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
Werkzeug==2.3.6
